server/requirements.txt
flask==1.0.2
flask-cors==3.0.9
paho-mqtt==1.5.1
python-rtmidi==1.4.6
flask-socketio==4.3.2
